gpu: Initial work on a Metal backend.
This hasn't even been compiled yet--I didn't write it on a Mac!--so this
likely has obvious syntax problems and copy/paste mistakes, etc.

This implements everything but the elephant in the room--shaders--since
I'm still deciding what I want to do there, so even once the thing
compiles you can't use it yet.

But this is a big chunk of progress!