gpu: Add Metal to list of usable drivers.
This actually works enough to run test/testgpu_simple_clear.c correctly
on a Mac, if you statically link SDL (since none of the GPU API entry
points are exported from the shared library or in the dynapi table yet),
which is surprisingly motivating for something that does so little.

Obviously we need to figure out the shader plan before anything else
is going to work, though.